+++ apply.go
@@ -24,7 +24,7 @@ func doApply(c *cli.Context) error {
 		return cli.NewExitError("`input-filename` is a required field.", 1)
 	}
 
-	cache, err := NewRedisCache()
+	cache, err := NewCache()
 	if err != nil {
 		return err
 	}
