+++ diagnosis.go
@@ -55,7 +55,7 @@ func doDiagnose(c *cli.Context) error {
 		return cli.NewExitError("`input-filename` is a required field.", 1)
 	}
 
-	cache, err := NewRedisCache()
+	cache, err := NewCache()
 	if err != nil {
 		return err
 	}
@@ -136,7 +136,7 @@ func doListFeatureWeight(c *cli.Context) error {
 		return cli.NewExitError("`input-filename` is a required field.", 1)
 	}
 
-	cache, err := NewRedisCache()
+	cache, err := NewCache()
 	if err != nil {
 		return err
 	}
